+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
+### Added
+- Image management to Badges. ([#92][i92])
 ### Fixed
 - Fix Admin dashboard render errors. ([#88][i88])
 - Fix deletion of the default level. ([#33][i33])
 
+[i92]: https://github.com/pacoorozco/gamify-laravel/issues/92
 [i88]: https://github.com/pacoorozco/gamify-laravel/issues/88
 [i33]: https://github.com/pacoorozco/gamify-laravel/issues/33