+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 ## Unreleased
 ### Added
 - AWS S3 support for image uploads.
-- Image management to Badges & Levels. ([#92][i92], [#96][i96]) 
+- Image management to Badges, Levels & Avatars. ([#92][i92], [#96][i96], [#99][i99]) 
 ### Fixed
 - Fix Admin dashboard render errors. ([#88][i88])
 - Fix deletion of the default level. ([#33][i33])
 
+[i99]: https://github.com/pacoorozco/gamify-laravel/issues/99
 [i96]: https://github.com/pacoorozco/gamify-laravel/issues/96
 [i92]: https://github.com/pacoorozco/gamify-laravel/issues/92
 [i88]: https://github.com/pacoorozco/gamify-laravel/issues/88